moves proposals and proposal new to pages
