+++ src/pages/Proposals/index.tsx
@@ -1,13 +1,13 @@
 import { Routes, Route, Navigate } from "react-router-dom";
 
-import New from "./New";
-import ProposalDetails from "./ProposalDetails";
+import ProposalDetails from "../../components/DAOs/DAO/Proposals/ProposalDetails";
+import ProposalCreate from "../ProposalCreate";
 
 function Proposals() {
   return (
     <Routes>
       <Route index element={<Navigate to="./.." replace={true} />} />
-      <Route path="new" element={<New />} />
+      <Route path="new" element={<ProposalCreate />} />
       <Route path=":proposalNumber/*" element={<ProposalDetails />} />
     </Routes>
   );
